test: address JavaScript permission review feedback
