+++ tests/functional/orchestration/javascript/workers/permissions_matrix_test.go
@@ -1,28 +1,15 @@
 package workers_test
 
 import (
-	"bytes"
-	"encoding/json"
-	"net/http"
+	"path/filepath"
 	"reflect"
-	"strings"
 	"testing"
 
 	serviceedges "github.com/portpowered/infinite-you/pkg/services/edges"
-	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	"github.com/portpowered/infinite-you/tests/functional/internal/support"
 )
 
 func TestJavaScriptAgentRunCodexCommandCharacterization(t *testing.T) {
-	dir := support.ScaffoldFactory(t, overridesFactoryConfig())
-	runner := support.NewRecordingCommandRunner("permission matrix child output")
-	server := support.StartFunctionalAPIServer(t, support.FunctionalAPIServerConfig{
-		FactoryDir:                dir,
-		WaitForServiceModeRuntime: true,
-		Edges:                     serviceedges.Edges{ProviderCommandRunner: runner},
-	})
-	t.Cleanup(func() { server.Stop(t) })
-
 	tests := []struct {
 		name            string
 		mode            string
@@ -64,30 +51,69 @@ func TestJavaScriptAgentRunCodexCommandCharacterization(t *testing.T) {
 		},
 	}
 
-	for index, test := range tests {
+	for _, test := range tests {
+		test := test
 		t.Run(test.name, func(t *testing.T) {
-			before := len(runner.Requests())
-			started := startPermissionMatrixWorkflow(
-				t,
-				server.URL(),
-				"javascript-permission-matrix-"+string(rune('a'+index)),
-				permissionMatrixWorkflow(test.skipPermissions),
-				test.mode,
-			)
-			if started.Status != factoryapi.FactorySessionDurableLifecycleStatusSucceeded {
-				t.Fatalf("session status = %q, want SUCCEEDED; result=%#v", started.Status, started.Result)
+			dir := support.ScaffoldFactory(t, permissionMatrixFactoryConfig(test.mode, permissionMatrixWorkflow(test.skipPermissions)))
+			runner := support.NewRecordingCommandRunner("permission matrix child output")
+			inputs := support.FakeInputs(t.Context(), []string{
+				"you", "--json", "run",
+				"--factory", filepath.Join(dir, "factory.json"),
+				"--output", "primary",
+				"--no-record",
+				"permission matrix prompt",
+			})
+			inputs.Input.WorkingDirectory = dir
+			homeDir := t.TempDir()
+			inputs.Input.Env = append(inputs.Input.Env, "HOME="+homeDir, "USERPROFILE="+homeDir)
+
+			if err := support.BuildProcess(t, serviceedges.Edges{
+				ProviderCommandRunner: runner,
+			}).Execute(inputs.Input); err != nil {
+				t.Fatalf("Process.Execute() error = %v\nstdout:\n%s\nstderr:\n%s", err, inputs.Stdout(), inputs.Stderr())
 			}
 
 			requests := runner.Requests()
-			if len(requests) != before+1 {
-				t.Fatalf("provider command requests = %d, want one new request; requests=%#v", len(requests)-before, requests)
+			if len(requests) != 1 {
+				t.Fatalf("provider command requests = %d, want one request; requests=%#v", len(requests), requests)
 			}
-			got := requests[before]
+			got := requests[0]
 			if got.Command != "codex" || !reflect.DeepEqual(got.Args, test.wantArgs) {
 				t.Fatalf("provider command = %q %#v, want codex %#v", got.Command, got.Args, test.wantArgs)
 			}
 		})
 	}
+}
+
+func permissionMatrixFactoryConfig(mode, source string) map[string]any {
+	defaultPolicy := map[string]any{}
+	if mode != "" {
+		defaultPolicy["mode"] = mode
+	}
+	config := map[string]any{}
+	config["name"] = "javascript-permission-matrix"
+	config["invocationSignature"] = map[string]any{
+		"parameters": []any{map[string]any{
+			"name": "prompt", "required": false,
+			"bindings": []any{map[string]any{"kind": "POSITIONAL", "position": 1}},
+		}},
+	}
+	config["orchestrator"] = map[string]any{
+		"kind": "JAVASCRIPT",
+		"javascript": map[string]any{
+			"inlineSource": map[string]any{
+				"encoding": "utf-8",
+				"inline":   source,
+			},
+			"argsSchema": map[string]any{
+				"type":                 "object",
+				"properties":           map[string]any{"prompt": map[string]any{"type": "string"}},
+				"additionalProperties": false,
+			},
+			"defaultPolicy": defaultPolicy,
+		},
+	}
+	return config
 }
 
 func permissionMatrixWorkflow(skipPermissions string) string {
@@ -102,59 +128,4 @@ func permissionMatrixWorkflow(skipPermissions string) string {
     modelProvider: "codex"` + field + `
   });
 })();`
-}
-
-func startPermissionMatrixWorkflow(
-	t *testing.T,
-	serverURL, requestID, workflowSource, mode string,
-) factoryapi.FactorySessionSyncExecutionResponse {
-	t.Helper()
-
-	dialect := "you-workflow-v1"
-	request := factoryapi.FactorySessionExecutionRequest{
-		RequestId: requestID,
-		Source: factoryapi.FactorySessionExecutionSource{
-			Kind: factoryapi.FactorySessionExecutionSourceKindInlineWorkflow,
-			InlineWorkflow: &factoryapi.FactorySessionExecutionInlineWorkflow{
-				Dialect: &dialect,
-				InlineSource: factoryapi.FactoryOrchestratorJavaScriptInlineSource{
-					Encoding: factoryapi.FactoryOrchestratorJavaScriptInlineSourceEncodingUtf8,
-					Inline:   workflowSource,
-				},
-			},
-		},
-	}
-	defaultPolicy := map[string]interface{}{}
-	if mode != "" {
-		defaultPolicy["mode"] = mode
-	}
-	request.Orchestrator = &factoryapi.FactoryOrchestrator{
-		Kind: factoryapi.JAVASCRIPT,
-		Javascript: &factoryapi.FactoryOrchestratorJavaScriptConfig{
-			DefaultPolicy: &defaultPolicy,
-		},
-	}
-	payload, err := json.Marshal(request)
-	if err != nil {
-		t.Fatalf("marshal permission matrix workflow request: %v", err)
-	}
-	endpoint := strings.TrimSuffix(serverURL, "/") + "/factory-sessions/sync"
-	httpRequest, err := http.NewRequestWithContext(t.Context(), http.MethodPost, endpoint, bytes.NewReader(payload))
-	if err != nil {
-		t.Fatalf("build permission matrix workflow request: %v", err)
-	}
-	httpRequest.Header.Set("Content-Type", "application/json")
-	response, err := http.DefaultClient.Do(httpRequest)
-	if err != nil {
-		t.Fatalf("start permission matrix workflow: %v", err)
-	}
-	defer response.Body.Close()
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("start permission matrix workflow status = %d", response.StatusCode)
-	}
-	var started factoryapi.FactorySessionSyncExecutionResponse
-	if err := json.NewDecoder(response.Body).Decode(&started); err != nil {
-		t.Fatalf("decode permission matrix workflow response: %v", err)
-	}
-	return started
 }
